feat: mock file and file with types
