+++ components/Feed.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import LeftSideBar from "./LeftSideBar";
 import Stories from "./Stories";
 import CreatePost from "./CreatePost";
@@ -6,15 +6,13 @@ import Posts from "./Posts";
 import RightSideBar from "./RightSideBar";
 
 const Feed = () => {
-  const [imageIsLoad, setImageIsLoad] = useState(false);
-
   return (
     <div className="flex bg-[#f2f3f7] flex-1 min-h-screen">
       <LeftSideBar />
       <div className="mx-auto ">
         <Stories />
-        <CreatePost /* onSetImageIsLoad={setImageIsLoad} */ />
-        <Posts onImageIsLoad={imageIsLoad} />
+        <CreatePost />
+        <Posts />
       </div>
       <RightSideBar />
     </div>
