+++ components/Header.tsx
@@ -1,16 +1,17 @@
 import Image from "next/image";
 import React from "react";
-import facebook from "../assets/facebook1.png";
+
 import { useSession, signIn, signOut } from "next-auth/react";
 import { useRouter } from "next/router";
 
+import facebook from "../assets/facebook1.png";
+import nouser from "../assets/nouser.png";
 import { MdHome } from "react-icons/md";
-import { FiPlayCircle, FiFlag, FiMessageCircle } from "react-icons/fi";
+import { FiPlayCircle, FiFlag } from "react-icons/fi";
 import { BsCart3 } from "react-icons/bs";
 import { GrGroup, GrAppsRounded } from "react-icons/gr";
 import { FaBell } from "react-icons/fa";
 import { AiOutlineMessage } from "react-icons/ai";
-import nouser from "../assets/nouser.png";
 
 const Header = () => {
   const { data: session } = useSession();
