+++ components/RightSideBar.tsx
@@ -1,37 +1,12 @@
 import React from "react";
-import Image, { StaticImageData } from "next/image";
+import Image from "next/image";
+import { profiles } from "../utils/mock";
 
-import albert from "../assets/1albert.jpg";
-import arnold from "../assets/1arnold.jpg";
-import drphil from "../assets/1drphil.webp";
-import elon from "../assets/1elon.webp";
-import kobe from "../assets/1kobe.webp";
-import miketyson from "../assets/1miketyson.jpg";
-import mrbeast from "../assets/1mrbeast.jpg";
-import rihana from "../assets/1rihana.jpg";
-import therock from "../assets/1rock.jpg";
-
+import dots from "../assets/dots.png";
 import { BsFillCameraVideoFill } from "react-icons/bs";
 import { FiSearch } from "react-icons/fi";
-import dots from "../assets/dots.png";
-
-type ProfilesType = {
-  name: string;
-  photo: StaticImageData;
-};
 
 const RightSideBar = () => {
-  const profiles: ProfilesType[] = [
-    { name: "Albert E.", photo: albert },
-    { name: "Arnold S.", photo: arnold },
-    { name: "Dr Phill", photo: drphil },
-    { name: "Elon Musk", photo: elon },
-    { name: "Kobe Briant", photo: kobe },
-    { name: "Mike Tyson", photo: miketyson },
-    { name: "Mr Beast", photo: mrbeast },
-    { name: "Rihana", photo: rihana },
-    { name: "The Rock", photo: therock },
-  ];
   return (
     <div>
       <div className="pt-4 sm:pt-8 pr-7 hidden lg:block">
